requirements.txt
opencv-python==4.8.1.78
ultralytics==8.0.196
streamlit==1.28.1
Pillow==10.0.1
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.2
seaborn==0.12.2
